Extract Tika and LLM backend clients from app
